# apps/backend/src/domains/auth/controllers/auth.controller.ts
// apps/backend/src/domains/auth/controllers/auth.controller.ts

import {
  Controller,
  Post,
  Get,
  Body,
  Req,
  Res,
  UseGuards,
  HttpCode,
  HttpStatus,
  UnauthorizedException,
  UseInterceptors,
  ClassSerializerInterceptor,
  Ip,
  Headers,
} from '@nestjs/common';
import { 
  ApiTags, 
  ApiOperation, 
  ApiResponse, 
  ApiBody, 
  ApiCookieAuth,
  ApiBearerAuth,
  ApiHeader
} from '@nestjs/swagger';
import { Request, Response } from 'express';
import { TenantContext, GetTenantContext } from '../../../shared/types/tenant-context';
import { AuthService } from '../services/auth.service';
import { TenantService } from '../../tenant/services/tenant.service';
import { LoginDto } from '../dto/login.dto';
import { LoginResponse } from '../interfaces/login-response.interface';
import { JwtAuthGuard, SkipAuth } from '../../../shared/guards';
import { AuthRateLimit, SkipRateLimit } from '../../../shared/decorators/multitenant-rate-limit.decorator';

@ApiTags('Authentication')
@Controller('auth')
export class AuthController {
  constructor(
    private readonly authService: AuthService,
    private readonly tenantService: TenantService,
  ) {}

  @Post('login')
  @ApiOperation({ 
    summary: 'Login to tenant',
    description: 'Authenticates user and returns JWT token. Can be used for both platform and tenant login.'
  })
  @ApiBody({ 
    type: LoginDto,
    description: 'User login credentials'
  })
  @ApiResponse({ 
    status: 200, 
    description: 'Login successful',
    schema: {
      type: 'object',
      properties: {
        accessToken: { type: 'string', description: 'JWT access token' }
      }
    }
  })
  @ApiResponse({ status: 401, description: 'Invalid credentials' })
  @ApiResponse({ status: 400, description: 'Invalid request data' })
  @ApiResponse({ status: 429, description: 'Too many login attempts' })
  @ApiHeader({
    name: 'x-forwarded-host',
    description: 'Tenant subdomain for tenant-specific login',
    required: false
  })
  @AuthRateLimit()
  @HttpCode(HttpStatus.OK)
  async login(
    @Body() dto: LoginDto,
    @GetTenantContext() tenant: { id: string; databaseUrl: string } | undefined,
    @Res({ passthrough: true }) res: Response,
  ): Promise<LoginResponse> {
    let tenantId: string | undefined = tenant?.id;
    
    // If tenantSubdomain is provided in the request body, resolve it to tenant ID
    if (dto.tenantSubdomain && !tenantId) {
      try {
        const tenantInfo = await this.tenantService.findBySubdomain(dto.tenantSubdomain);
        tenantId = tenantInfo.id;
        console.log(`[AUTH] Resolved tenant subdomain "${dto.tenantSubdomain}" to ID: ${tenantId}`);
      } catch (error) {
        console.error(`[AUTH] Failed to resolve tenant subdomain "${dto.tenantSubdomain}":`, error);
        throw new UnauthorizedException(`Invalid tenant: ${dto.tenantSubdomain}`);
      }
    }
    
    console.log(`[AUTH] Login attempt - Email: ${dto.email}, TenantID: ${tenantId || 'platform'}`);
    const { accessToken } = await this.authService.login(dto, tenantId);

    res.cookie('Authentication', accessToken, {
      httpOnly: true,
      secure: process.env.NODE_ENV === 'production',
      sameSite: 'strict',
      maxAge: 1000 * 60 * 60, // 1h in ms
    });

    return { accessToken };
  }

  @Get('me')
  @ApiOperation({ 
    summary: 'Get current user',
    description: 'Returns the currently authenticated user information.'
  })
  @ApiCookieAuth('Authentication')
  @ApiBearerAuth('JWT-auth')
  @ApiResponse({ 
    status: 200, 
    description: 'Current user information',
    schema: {
      type: 'object',
      properties: {
        id: { type: 'string' },
        email: { type: 'string' },
        firstName: { type: 'string' },
        lastName: { type: 'string' },
        role: { type: 'string' },
        tenantId: { type: 'string', nullable: true }
      }
    }
  })
  @ApiResponse({ status: 401, description: 'Unauthorized' })
  @SkipRateLimit() // /me endpoint used frequently for auth verification
  @UseGuards(JwtAuthGuard)
  @HttpCode(HttpStatus.OK)
  me(@Req() req: Request) {
    const user = req.user;
    if (!user) {
      throw new UnauthorizedException();
    }
    return user;
  }

  @Get('csrf-token')
  @ApiOperation({ 
    summary: 'Get CSRF token',
    description: 'Returns a CSRF token for secure form submissions.'
  })
  @ApiResponse({ 
    status: 200, 
    description: 'CSRF token',
    schema: {
      type: 'object',
      properties: {
        csrfToken: { type: 'string', nullable: true }
      }
    }
  })
  @SkipRateLimit() // CSRF tokens are fetched frequently for security
  @HttpCode(HttpStatus.OK)
  getCsrfToken(@Req() req: Request, @Res({ passthrough: true }) res: Response) {
    // The CSRF token is generated by the middleware and available as req.csrfToken()
    const token = (req as any).csrfToken?.();
    if (token) {
      res.setHeader('X-CSRF-Token', token);
      return { csrfToken: token };
    }
    return { csrfToken: null };
  }

  @Post('logout')
  @ApiOperation({ 
    summary: 'Logout user',
    description: 'Logs out the current user and clears the authentication cookie.'
  })
  @ApiResponse({ 
    status: 200, 
    description: 'Logout successful',
    schema: {
      type: 'object',
      properties: {
        success: { type: 'boolean' }
      }
    }
  })
  @ApiResponse({ status: 429, description: 'Too many logout attempts' })
  @AuthRateLimit()
  @HttpCode(HttpStatus.OK)
  async logout(@Res({ passthrough: true }) res: Response) {
    const baseDomain = process.env.BASE_DOMAIN || 'lvh.me';
    // Log for debugging
    console.log(`Clearing Authentication cookie for .${baseDomain}, ${baseDomain}, /`);
    res.clearCookie('Authentication', {
      httpOnly: true,
      secure: process.env.NODE_ENV === 'production',
      sameSite: 'strict',
      domain: `.${baseDomain}`,
      path: '/',
    });
    res.clearCookie('Authentication', {
      httpOnly: true,
      secure: process.env.NODE_ENV === 'production',
      sameSite: 'strict',
      domain: baseDomain,
      path: '/',
    });
    res.clearCookie('Authentication', {
      httpOnly: true,
      secure: process.env.NODE_ENV === 'production',
      sameSite: 'strict',
      path: '/',
    });
    res.clearCookie('Authentication');
    return { success: true };
  }
}